LayoutTests/plugins/npruntime/evaluate.html


<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<script>
function runTest() {
    plugin = document.getElementById('plugin');

    shouldBe("plugin.testEvaluate('1')", "1");
    shouldBe("plugin.testEvaluate('plugin')", "plugin");
    shouldBe("plugin.testEvaluate('window')", "window");
    shouldBe("plugin.testEvaluate('1 + 1')", "2");
}
</script>
<body onLoad="runTest()">
<embed id="plugin" type="application/x-webkit-test-netscape" width=100 height=100></embed>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test NPN_Evaluate");

</script>
